Many changes: - Improved Rhino conversion code by adding getConversionWeight to WrapFactory. - Improved Rhino Java constructor handling with properties / functions by looking at last parameter of constructors and argument lists. - Changed MenuGroup / MenuItem to auto-generate names and not receive name / text values in the constructor (to reflect the way all the other ADM objects are created). - Introduced com.scriptographer.ai.Size and let Document use it. - Improved  getConversionWeight code in RhinoWrapFactory, e.g. to distinguish setSize versions that receive both Point and Size, so it prefers ai.Size over a Point when setting adm.Size. - Improved GUI code. - Added code to ListWrapper to check for get/setSize when getting / setting "size", in order to mask the native size() function, since "length" returns that already, and e.g. HierarchyList has the set/getSize methods for its dimensions. - Updated to the latest modified Rhino version. - Added a patch for Scriptographer's changes to Rhino.

+++ src/java/com/scriptographer/ai/Rectangle.java
@@ -44,7 +44,6 @@ import com.scratchdisk.util.ConversionUtils;
  * @author lehni
  */
 public class Rectangle {
-
 	protected float x;
 	protected float y;
 	protected float width;
@@ -106,10 +105,10 @@ public class Rectangle {
 	 * @param map <code>{x, y, width, height}</code>
 	 */
 	public Rectangle(Map map) {
-		this(ConversionUtils.getDouble(map, "x"),
-				ConversionUtils.getDouble(map, "y"),
-				ConversionUtils.getDouble(map, "width"),
-				ConversionUtils.getDouble(map, "height"));
+		this(ConversionUtils.getFloat(map, "x"),
+				ConversionUtils.getFloat(map, "y"),
+				ConversionUtils.getFloat(map, "width"),
+				ConversionUtils.getFloat(map, "height"));
 	}
 
 	/**
@@ -287,7 +286,7 @@ public class Rectangle {
 		setRight(x);
 		setBottom(y);
 	}
-	
+
 	public String toString() {
 		StringBuffer buf = new StringBuffer(128);
 		buf.append("{ x: ").append(x);
@@ -297,7 +296,7 @@ public class Rectangle {
 		buf.append(" }");
 		return buf.toString();
 	}
-	
+
 	/**
 	 * Clones the rectangle.
 	 */
